+++ src/EntityFrameworkCore.Triggered/Internal/CascadingStrategies/EntityAndTypeCascadingStrategy.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore.ChangeTracking;
 
-namespace EntityFrameworkCore.Triggered.Internal.CascadeStrategies
+namespace EntityFrameworkCore.Triggered.Internal.CascadingStrategies
 {
-    public class EntityAndTypeCascadeStrategy : ICascadeStrategy
+    public class EntityAndTypeCascadingStrategy : ICascadingStrategy
     {
         public bool CanCascade(EntityEntry entry, ChangeType changeType, TriggerContextDescriptor previousTriggerContextDescriptor)
             => changeType != previousTriggerContextDescriptor.ChangeType;
