+++ src/EntityFrameworkCore.Triggered/Internal/CascadingStrategies/ICascadingStrategy.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore.ChangeTracking;
 
-namespace EntityFrameworkCore.Triggered.Internal.CascadeStrategies
+namespace EntityFrameworkCore.Triggered.Internal.CascadingStrategies
 {
-    public interface ICascadeStrategy
+    public interface ICascadingStrategy
     {
         bool CanCascade(EntityEntry entry, ChangeType changeType, TriggerContextDescriptor previousTriggerContextDescriptor);
     }
