+++ src/EntityFrameworkCore.Triggered/Internal/CascadingStrategies/NoCascadingStrategy.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore.ChangeTracking;
 
-namespace EntityFrameworkCore.Triggered.Internal.CascadeStrategies
+namespace EntityFrameworkCore.Triggered.Internal.CascadingStrategies
 {
-    public class NoCascadeStrategy : ICascadeStrategy
+    public class NoCascadingStrategy : ICascadingStrategy
     {
         public bool CanCascade(EntityEntry entry, ChangeType changeType, TriggerContextDescriptor previousTriggerContextDescriptor)
             => false;
