+++ test/EntityFrameworkCore.Triggered.Tests/Internal/CascadingStrategies/RecursionStrategyTestsBase.cs
@@ -1,5 +1,5 @@
 ﻿using EntityFrameworkCore.Triggered.Internal;
-using EntityFrameworkCore.Triggered.Internal.CascadeStrategies;
+using EntityFrameworkCore.Triggered.Internal.CascadingStrategies;
 using Microsoft.EntityFrameworkCore;
 using Xunit;
 
@@ -20,7 +20,7 @@ namespace EntityFrameworkCore.Triggered.Tests.Internal.CascadeStrategies
             }
         }
 
-        protected abstract ICascadeStrategy CreateSubject();
+        protected abstract ICascadingStrategy CreateSubject();
 
         protected abstract bool CanCascadeUnmodifiedExpectedOutcome { get; }
         protected abstract bool CanCascadeModifiedExpectedOutcome { get; }
